icc_utils: cleanup for buildroot
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@10684 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ user/blkfin-apps/icc_utils/include/mempool.h
@@ -32,9 +32,9 @@ struct gen_pool_chunk {
 };
 
 extern struct gen_pool *gen_pool_create(int);
-extern int gen_pool_add(struct gen_pool *, unsigned long, size_t);
+extern int gen_pool_add(struct gen_pool *, unsigned long, unsigned long);
 extern void gen_pool_destroy(struct gen_pool *);
-extern unsigned long gen_pool_alloc(struct gen_pool *, size_t);
-extern void gen_pool_free(struct gen_pool *, unsigned long, size_t);
+extern unsigned long gen_pool_alloc(struct gen_pool *, unsigned long);
+extern void gen_pool_free(struct gen_pool *, unsigned long, unsigned long);
 
 #endif
